==> _episodes_rmd/13-visualizing-relationships.Rmd <==
---
title: "Building a regression model"
teaching: 45
exercises: 10
questions:
- "How can I perform a univariate regression (one x variable)?"
- "How do I interpret the results of a regression?"
- "How can I make a graph of my regression?"
- "How can I look at residuals?"
- "How would I perform a logistic regresion?"
- "How can I build a multivariate model and perform regression?"
- "How can I perform model selection?"
- "How can I assess and remove outliers?"
objectives:
- TODO: These need updating
- "Make a scatterplot"
- "`cor()`"
- "Learn how to run regressions with `lm()` and `glm()`"
- "Learn how to isolate parts of the matrix that `lm()` returns"
- "Learn how to run `t.test()` - get p-value, CI, etc."
- "Assess interactions between variables"
keypoints:
- "FIXME"
source: Rmd
---

We are finally ready to begin a full (multivariate) linear regression model. To do this, we are going to use information for the univariate models in the last episode.  

```{r, include=FALSE}
source("../bin/chunk-options.R")
knitr_fig_path("13-")
```

```{r, include=FALSE}
library(dplyr)

analysis_swan_df <- readRDS(file = "data_out/analysis_swan_df_12.rds")
colnames(analysis_swan_df)
```


# Remove rows with NAs in any of the following variables:
```{r}
library(tidyr)

analysis_swan_df <- analysis_swan_df %>% drop_na(Glucose, BMI_cat, bp_category, Age, Chol_Ratio,
                                                 RACE, log_CRP, Smoker, Exercise, age_Chol_Ratio)
```

```{r}
full_model <- lm(Glucose ~ BMI_cat + bp_category  + Age + Chol_Ratio +
                 RACE + log_CRP + Smoker + Exercise + age_Chol_Ratio,
                 data = analysis_swan_df)
summary(full_model)
```

#TODO Create a stepwise selection and see what stays in the model. Based on that we will do resids etc. 

Plotting residuals
```{r}
plot(resid(full_model))
```

```{r}
rst <- rstandard(full_model)
qqnorm(rst, ylab = "Standardized Residuals",
       xlab = "Normal Scores")
qqline(rst)
```


# Stepwise Regression

```{r}
library(MASS)
# Stepwise regression model
step.model <- stepAIC(full_model, direction = "backward",
                      trace = TRUE )
summary(step.model)
```
```{r}
min_model <- lm(Glucose ~ BMI_cat,
                data = analysis_swan_df)

step.model <- stepAIC(min_model, direction = "forward",
                      scope = list(lower = min_model, upper = full_model),
                      trace = TRUE )

summary(step.model)
```


```{r}
analysis_swan_df$cooksd <- cooks.distance(step.model)
cooksd_cutoff <- 4/nrow(analysis_swan_df)

plot(analysis_swan_df$cooksd)
abline(cooksd_cutoff, 0)
```

Dropping outliers
```{r}
analysis_swan_df2 <- analysis_swan_df %>% filter(cooksd < cooksd_cutoff) %>% dplyr::select(-cooksd)

step.model.2  <- lm(Glucose ~ BMI_cat + bp_category + RACE + Exercise + log_CRP + age_Chol_Ratio,
                    data = analysis_swan_df2)
summary(step.model.2)
```

### Note big increase in Adj. R-squared. 


Plotting residuals
```{r}
plot(resid(step.model.2))
```

```{r}
rst <- rstandard(step.model.2)
qqnorm(rst, ylab = "Standardized Residuals",
       xlab = "Normal Scores")
qqline(rst)
```

```{r}
analysis_swan_lt_126 <- analysis_swan_df2 %>% filter(Glucose < 126)

step.model.3  <- lm(Glucose ~ BMI_cat + bp_category + RACE + Exercise + log_CRP + age_Chol_Ratio,
                    data = analysis_swan_lt_126)
summary(step.model.3)
```

```{r}
plot(resid(step.model.3))
```

```{r}
rst <- rstandard(step.model.3)
qqnorm(rst, ylab = "Standardized Residuals",
       xlab = "Normal Scores")
qqline(rst)
```


### VIF - Collinearity


```{r}

library(car)

vif(step.model.2)
# VIF has to be <=1 ???

#pairs(analysis_swan_df2 %>% dplyr::select(AGE6,  HDL_LDL))
```
```{r}
vif(full_model)
```

```{r}
#pairs(analysis_swan_df2 %>% dplyr::select(Age, Chol_Ratio, age_Chol_Ratio))
```
## TODO: Durbin-Watson test for autocorrelation in residuals

```{r}
library(car)
car::dwt(step.model)
```
